requirements.txt
requests>=2.25.1
python-dotenv>=0.21.0
python-telegram-bot>=20.0
lyricsgenius>=1.8.0
Flask>=2.0
